+++ test/recipes/70-test_sslsessiontick.t
@@ -24,8 +24,8 @@ plan skip_all => "$test_name needs the dynamic engine feature enabled"
 plan skip_all => "$test_name needs the sock feature enabled"
     if disabled("sock");
 
-plan skip_all => "$test_name needs SSLv3, TLSv1, TLSv1.1 or TLSv1.2 enabled"
-    if alldisabled(("ssl3", "tls1", "tls1_1", "tls1_2"));
+plan skip_all => "$test_name needs TLSv1, TLSv1.1 or TLSv1.2 enabled"
+    if alldisabled(("tls1", "tls1_1", "tls1_2"));
 
 sub checkmessages($$$$$$);
 sub clearclient();